backend/requirements.txt
# IdentiFace Studio - Backend Dependencies

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
databases==0.8.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Face Recognition & Computer Vision
face-recognition==1.3.0
dlib==19.24.2
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
scikit-image==0.22.0

# Image Processing
scipy==1.11.4
mtcnn==0.1.1

# Utilities
aiofiles==23.2.1
python-dateutil==2.8.2
email-validator==2.1.0

# Caching
redis==5.0.1
hiredis==2.2.3

# Monitoring & Logging
loguru==0.7.2
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# CORS
fastapi-cors==0.0.6

# File handling
aiofiles==23.2.1

# Background tasks
celery==5.3.4
